main and user class
Main is functional. It is still very bare bones as of now. It currently supports the creation of an account, logging in, logging out, and viewing all userIDs in the database.
User class is functional. It supports all of the mechanics listed above, but a couple of improvements still need to be made.